add owner to  contact methods

+++ controllers/contacts/removeById.js
@@ -4,7 +4,7 @@ const { successResponse } = require('../../helpers')
 
 const removeById = async (req, res) => {
   const { id } = req.params
-  const result = await Contact.findByIdAndRemove(id)
+  const result = await Contact.findByIdAndRemove({ owner: id })
   if (!result) {
     throw new NotFound(`Contact with id ${id} not found`)
   }
